moved home and chat screen files to respective dir
